Refactor AGVController to improve logging of AGV locations and support multiple AGVs, and update AGV movement logic

+++ Assets/Scripts/Floor Plan/AssignCordinateMapping.cs
@@ -30,8 +30,8 @@ public class AutoLabelGrandChildren : MonoBehaviour
                     // int X = (int)((49 - position.z) / 2);
                     // int Y = (int)((position.x + 5) / 2);
                     //Debug.Log(position);
-                    int X = (int)((((int)position.z) + 22 )/2);   // equation is x+5/2
-                    int Y = (int)(((49- (int)position.x)+6)/2);     // equation is 49-x/2
+                    int X = (int)(((position.z) + 22 )/2);   // equation is x+5/2
+                    int Y = (int)(((49- position.x)+6)/2);     // equation is 49-x/2
 
                     // Set the label text to the sphere's coordinates
                     label.text = $"({X}, {Y})";
